# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Streambox API
  description: API REST para la gestión de películas, géneros y usuarios de Streambox.
  version: '1.0'
servers:
- url: http://localhost:8080
  description: Backend local (perfil dev)
paths:
...
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      summary: Renueva la sesión
      description: 'Entrega un JWT de acceso nuevo (cookie streambox_token) a partir del refresh token de la cookie streambox_refresh,
        sin pedir la contraseña. Sin cuerpo. Solo lee la cookie streambox_refresh: ignora streambox_token y la cabecera Authorization.
        Exige la cabecera X-Requested-With: StreamBox (defensa CSRF). Cada uso rota el refresh token: el viejo queda revocado
        y llega uno nuevo en Set-Cookie (vida de 7 días, sin pasar de 30 días desde el login). Reutilizar un refresh token
        ya rotado se trata como un robo y revoca la sesión entera (también el token nuevo), salvo en los 10 segundos siguientes
        a la rotación: así dos pestañas que renuevan a la vez no se cierran la sesión; la segunda recibe solo el JWT nuevo,
        sin Set-Cookie de streambox_refresh, porque el navegador ya tiene el sucesor.'
      operationId: refresh
      parameters:
      - name: streambox_refresh
        in: cookie
        description: Refresh token (cookie HttpOnly streambox_refresh que fija el login; el navegador la envía solo)
        required: false
        schema:
          type: string
      responses:
        '204':
          description: Sesión renovada; sin cuerpo. Set-Cookie con streambox_token y, salvo en la gracia de un refresh simultáneo,
            con el streambox_refresh nuevo
        '401':
          description: 'Código SESSION_EXPIRED: el refresh token falta, no existe, ha caducado (él o su sesión de 30 días),
            está revocado (logout) o se ha reutilizado (y entonces se revoca toda la sesión). Mismo cuerpo en todos los casos.
            La respuesta borra las dos cookies: hay que volver a iniciar sesión'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Código CSRF_REJECTED: falta la cabecera X-Requested-With: StreamBox. No se mira el refresh token ni
            se gasta el límite por IP'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Código RATE_LIMIT_EXCEEDED: demasiadas renovaciones desde esta IP (30 por minuto por defecto); la
            cabecera Retry-After indica los segundos que hay que esperar. Solo cuentan las peticiones con la cabecera X-Requested-With
            y con la cookie streambox_refresh: sin cookie (visita sin sesión) la respuesta es siempre el 401 SESSION_EXPIRED'
          headers:
            Retry-After:
              description: Segundos que hay que esperar antes de volver a intentarlo (entero, mínimo 1).
              required: true
              schema:
                type: integer
                format: int32
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      summary: Cierra sesión
      description: 'Revoca en el servidor la sesión del refresh token de la cookie streambox_refresh (sus tokens dejan de
        servir para renovar) y borra las cookies streambox_token y streambox_refresh (Set-Cookie con Max-Age=0). Es público
        e idempotente: responde 204 haya o no sesión, y con un refresh token desconocido o ya revocado. Exige la cabecera
        X-Requested-With: StreamBox (defensa CSRF, también con Authorization: Bearer): sin ella otra web podría borrar las
        cookies de la víctima. El JWT de acceso es stateless y no se puede revocar: una copia seguiría valiendo hasta su caducidad
        (como mucho 15 minutos por defecto), pero ya no se podría renovar.'
      operationId: logout
      parameters:
      - name: streambox_refresh
        in: cookie
        description: Refresh token de la sesión que se cierra (cookie HttpOnly streambox_refresh); opcional
        required: false
        schema:
          type: string
      responses:
        '204':
          description: Sesión revocada (si la había) y cookies borradas
        '403':
          description: 'Código CSRF_REJECTED: falta la cabecera X-Requested-With: StreamBox. No se revoca nada ni se borran
            las cookies: la sesión sigue abierta'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Código INTERNAL_ERROR: no se ha podido revocar la sesión (por ejemplo, la base de datos no responde).
            No se borran las cookies: la sesión sigue viva en el servidor y el cliente debe reintentar el logout en lugar
            de darla por cerrada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout-all:
    post:
      tags:
      - auth-controller
      summary: Cierra sesión en todos los dispositivos
      description: 'Revoca todas las sesiones de la cuenta autenticada, incluida la actual: ningún refresh token suyo vuelve
        a servir para renovar. Borra las cookies streambox_token y streambox_refresh de esta petición (Set-Cookie con Max-Age=0).
        Sin cuerpo. Los JWT de acceso ya emitidos son stateless y no se pueden revocar: siguen valiendo hasta caducar (como
        mucho 15 minutos), pero ya no se pueden renovar.'
      operationId: logoutAll
      responses:
        '204':
          description: Todas las sesiones revocadas y cookies borradas
        '401':
          description: El usuario no está autenticado (no se revoca ni se borra nada)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Código CSRF_REJECTED: la petición se autentica por la cookie streambox_token y le falta la cabecera
            X-Requested-With: StreamBox (defensa CSRF). Con Authorization: Bearer no se exige'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Código INTERNAL_ERROR: no se han podido revocar las sesiones (por ejemplo, la base de datos no responde).
            No se borran las cookies: el cliente debe reintentar en lugar de dar las sesiones por cerradas'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
      - bearerAuth: []
  /api/auth/login:
    post:
      tags:
      - auth-controller
      summary: Inicia sesión
      description: 'Autentica un usuario mediante su correo electrónico y contraseña. Los tokens no van en el cuerpo: se entregan
        en dos cookies HttpOnly; SameSite=Strict (Secure según la configuración). streambox_token: JWT de acceso, Path=/api,
        Max-Age = su vida (15 minutos por defecto). streambox_refresh: refresh token opaco, Path=/api/auth (solo viaja a login,
        refresh y logout), Max-Age = su vida (7 días). Cuando el JWT caduca, POST /api/auth/refresh entrega uno nuevo sin
        volver a pedir la contraseña. Las peticiones POST/PUT/PATCH/DELETE autenticadas por cookie deben llevar la cabecera
        X-Requested-With: StreamBox (403 CSRF_REJECTED si falta). También se acepta Authorization: Bearer con el JWT de acceso
        (clientes de API), sin esa cabecera extra.'
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '204':
          description: Autenticación correcta; sin cuerpo, con las cookies streambox_token y streambox_refresh en Set-Cookie
        '400':
          description: 'Los datos proporcionados no son válidos (email con formato incorrecto, contraseña vacía o de más de
            1024 caracteres). No gasta intento de la cuenta. En el login no se aplica la política de longitud del registro:
            las cuentas antiguas siguen pudiendo entrar'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Credenciales incorrectas (código INVALID_CREDENTIALS, mensaje «Email o contraseña incorrectos»). El
            cuerpo incluye remainingAttempts: intentos que le quedan a la cuenta ya descontado este fallo (siempre 1 o más);
            si el siguiente intento también falla y era el último, la cuenta se bloquea y se responde 429 ACCOUNT_LOCKED.
            Desde una misma IP de origen es idéntico exista o no el email'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 'Demasiados intentos; la cabecera Retry-After indica los segundos que hay que esperar. Código ACCOUNT_LOCKED:
            la cuenta está bloqueada por logins fallidos (lo devuelve el fallo que agota los intentos y cualquier intento
            durante el bloqueo, aunque la contraseña sea correcta; el mensaje dice cuántos minutos dura; no se aplica a una
            IP desde la que el titular ya inició sesión con éxito, que tiene su propio contador con los mismos límites). Código
            RATE_LIMIT_EXCEEDED: demasiadas peticiones de login desde esta IP'
          headers:
            Retry-After:
              description: Segundos que hay que esperar antes de volver a intentarlo (entero, mínimo 1).
              required: true
              schema:
                type: integer
                format: int32
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          maxLength: 1024
          minLength: 0
      required:
      - email
      - password
...
    ErrorResponse:
      type: object
      description: 'Formato común de todas las respuestas de error de la API (4xx y 5xx). El cliente debe decidir por status
        y code, nunca por el texto de message. Los campos que no aplican a un error se omiten del JSON: nunca se envían con
        valor null.'
      properties:
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora del error, en UTC y formato ISO-8601.
          example: '2026-10-04T11:15:30.123Z'
        status:
          type: integer
          format: int32
          description: Código de estado HTTP; coincide con el de la respuesta.
          example: 404
        error:
          type: string
          description: Frase estándar del estado HTTP según la especificación (en inglés).
          example: Not Found
        code:
          type: string
          description: 'Código funcional estable del error: el cliente debe decidir por este campo (y por status), nunca por
            el texto de message. Valores:

            - RESOURCE_NOT_FOUND (404): el recurso o la ruta no existen (una serie sin episodios también da este código a
            los usuarios).

            - MOVIE_NOT_IN_FAVORITES (404): la película no está en la lista del usuario.

            - SERIES_NOT_IN_FAVORITES (404): la serie existe, pero no está en la lista del usuario.

            - CSRF_REJECTED (403): petición no segura (POST/PUT/PATCH/DELETE) autenticada por cookie sin la cabecera X-Requested-With:
            StreamBox. POST /api/auth/refresh y POST /api/auth/logout la exigen siempre, también con Authorization: Bearer.

            - USER_ALREADY_EXISTS (409): el nombre de usuario o el email ya están registrados.

            - MOVIE_ALREADY_IN_FAVORITES (409): la película ya está en la lista del usuario.

            - SERIES_ALREADY_IN_FAVORITES (409): la serie ya está en la lista del usuario.

            - EPISODE_ALREADY_EXISTS (409): la serie ya tiene un episodio con esa temporada y número.

            - GENRE_ALREADY_EXISTS (409): ya existe otro género con ese nombre.

            - GENRE_IN_USE (409): el género no se puede borrar porque alguna película o serie lo tiene asignado.

            - AMBIGUOUS_TITLE (409): varias películas tienen ese título; hay que usar el id.

            - DATA_INTEGRITY_VIOLATION (409): la operación choca con una restricción de los datos.

            - VALIDATION_ERROR (400): datos o parámetros no válidos; el detalle por campo va en validationErrors.

            - CURRENT_PASSWORD_INCORRECT (400): al cambiar la contraseña, la actual no es correcta; el detalle va en validationErrors.currentPassword.
            No es un 401: la sesión sigue siendo válida.

            - MALFORMED_REQUEST (400 y otros 4xx genéricos): la petición no se puede interpretar (JSON mal formado, tipos
            incorrectos).

            - INVALID_CREDENTIALS (401): email o contraseña incorrectos en el login, o token ausente, caducado o no válido.

            - SESSION_EXPIRED (401): POST /api/auth/refresh no puede renovar la sesión (refresh token ausente, caducado, revocado
            o reutilizado); hay que volver a iniciar sesión.

            - ACCESS_DENIED (403): autenticado, pero sin permisos (por ejemplo, no es ADMIN).

            - METHOD_NOT_ALLOWED (405): el método HTTP no está soportado por la ruta.

            - UNSUPPORTED_MEDIA_TYPE (415): el tipo de contenido no está soportado.

            - NOT_ACCEPTABLE (406): la cabecera Accept pide un formato que la API no produce; la API solo responde en JSON
            (este error también va en JSON).

            - RATE_LIMIT_EXCEEDED (429): demasiadas peticiones de login, registro o renovación de sesión (POST /api/auth/refresh)
            desde la misma IP, o demasiados intentos con la contraseña actual incorrecta en PUT /api/users/me/password (por
            cuenta: 5 cada 15 minutos); ver la cabecera Retry-After.

            - ACCOUNT_LOCKED (429): la cuenta está bloqueada temporalmente por logins fallidos; ver la cabecera Retry-After.

            - INTERNAL_ERROR (500 u otro 5xx): error no controlado en el servidor.'
          enum:
          - RESOURCE_NOT_FOUND
          - USER_ALREADY_EXISTS
          - MOVIE_ALREADY_IN_FAVORITES
          - MOVIE_NOT_IN_FAVORITES
          - VALIDATION_ERROR
          - INVALID_CREDENTIALS
          - DATA_INTEGRITY_VIOLATION
          - INTERNAL_ERROR
          - AMBIGUOUS_TITLE
          - ACCESS_DENIED
          - MALFORMED_REQUEST
          - METHOD_NOT_ALLOWED
          - UNSUPPORTED_MEDIA_TYPE
          - NOT_ACCEPTABLE
          - RATE_LIMIT_EXCEEDED
          - ACCOUNT_LOCKED
          - GENRE_ALREADY_EXISTS
          - GENRE_IN_USE
          - EPISODE_ALREADY_EXISTS
          - SERIES_ALREADY_IN_FAVORITES
          - SERIES_NOT_IN_FAVORITES
          - CSRF_REJECTED
          - SESSION_EXPIRED
          - CURRENT_PASSWORD_INCORRECT
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: 'Mensaje legible en español, pensado para mostrarlo al usuario. Puede cambiar: no se debe usar para
            tomar decisiones.'
          example: Película no encontrada
        path:
          type: string
          description: Ruta de la petición que produjo el error (sin los parámetros de consulta).
          example: /api/movies/42
        validationErrors:
          type: object
          additionalProperties:
            type: string
          description: 'Solo en los 400 con código VALIDATION_ERROR o CURRENT_PASSWORD_INCORRECT: un mensaje por cada campo
            o parámetro no válido (nombre del campo → mensaje). En el resto de errores no aparece (no se envía con null).'
          example:
            title: El título es obligatorio
        remainingAttempts:
          type: integer
          format: int32
          description: 'Solo en el 401 INVALID_CREDENTIALS del login: intentos que le quedan a la cuenta, con este fallo ya
            descontado. Nunca vale 0, porque el fallo que agota los intentos ya responde 429 ACCOUNT_LOCKED. En el resto de
            errores no aparece (no se envía con null).'
          example: 4
          minimum: 1
      required:
      - code
      - error
      - message
      - path
      - status
      - timestamp
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: 'Autenticación principal (navegador): el JWT de acceso viaja en la cookie HttpOnly streambox_token que
        fija POST /api/auth/login (Path=/api, 15 minutos; se renueva con POST /api/auth/refresh). Las peticiones no seguras
        (POST, PUT, PATCH, DELETE) autenticadas por esta cookie deben llevar además la cabecera X-Requested-With: StreamBox
        (defensa CSRF); si falta, 403 CSRF_REJECTED. Si la petición trae Authorization: Bearer, manda el Bearer y la cookie
        no se mira. Swagger UI no puede fijar esta cookie: para probar desde aquí usa bearerAuth.'
      name: streambox_token
      in: cookie
    bearerAuth:
      type: http
      description: 'Alternativa para clientes de API sin navegador: el mismo JWT de acceso (el valor de la cookie streambox_token)
        en la cabecera Authorization: Bearer. Con Bearer no se exige X-Requested-With (no hay riesgo de CSRF).'
      scheme: bearer
      bearerFormat: JWT